+++ src/rules/no-unregistered-classes.ts
@@ -1,4 +1,3 @@
-import { getUnregisteredClasses } from "better-tailwindcss:async/unregistered-classes.sync.js";
 import {
   DEFAULT_ATTRIBUTE_NAMES,
   DEFAULT_CALLEE_NAMES,
@@ -13,11 +12,15 @@ import {
   TAILWIND_CONFIG_SCHEMA,
   VARIABLE_SCHEMA
 } from "better-tailwindcss:options/descriptions.js";
+import { getCustomComponentClasses } from "better-tailwindcss:tailwindcss/custom-component-classes.js";
+import { getPrefix } from "better-tailwindcss:tailwindcss/prefix.js";
+import { getUnregisteredClasses } from "better-tailwindcss:tailwindcss/unregistered-classes.js";
 import { getCommonOptions } from "better-tailwindcss:utils/options.js";
 import { createRuleListener } from "better-tailwindcss:utils/rule.js";
 import {
   augmentMessageWithWarnings,
   display,
+  escapeForRegex,
   getExactClassLocation,
   splitClasses
 } from "better-tailwindcss:utils/utils.js";
@@ -41,6 +44,7 @@ export type Options = [
     TagOption &
     VariableOption &
     {
+      detectComponentClasses?: boolean;
       entryPoint?: string;
       ignore?: string[];
       tailwindConfig?: string;
@@ -48,14 +52,12 @@ export type Options = [
   >
 ];
 
-export const DEFAULT_IGNORED_UNREGISTERED_CLASSES = [
-  "^group(?:\\/(\\S*))?$",
-  "^peer(?:\\/(\\S*))?$"
-];
+export const DEFAULT_IGNORED_UNREGISTERED_CLASSES = [];
 
 const defaultOptions = {
   attributes: DEFAULT_ATTRIBUTE_NAMES,
   callees: DEFAULT_CALLEE_NAMES,
+  detectComponentClasses: false,
   ignore: DEFAULT_IGNORED_UNREGISTERED_CLASSES,
   tags: DEFAULT_TAG_NAMES,
   variables: DEFAULT_VARIABLE_NAMES
@@ -84,6 +86,11 @@ export const noUnregisteredClasses: ESLintRule<Options> = {
             ...TAG_SCHEMA,
             ...ENTRYPOINT_SCHEMA,
             ...TAILWIND_CONFIG_SCHEMA,
+            detectComponentClasses: {
+              default: defaultOptions.detectComponentClasses,
+              description: "Whether to automatically detect custom component classes from the tailwindcss config.",
+              type: "boolean"
+            },
             ignore: {
               description: "A list of classes that should be ignored by the rule.",
               items: {
@@ -101,23 +108,34 @@ export const noUnregisteredClasses: ESLintRule<Options> = {
 };
 
 function lintLiterals(ctx: Rule.RuleContext, literals: Literal[]) {
+  const { detectComponentClasses, ignore, tailwindConfig } = getOptions(ctx);
+
+  const { prefix, suffix } = getPrefix({ configPath: tailwindConfig, cwd: ctx.cwd });
+
+  const ignoredGroups = new RegExp(`^${escapeForRegex(`${prefix}${suffix}`)}group(?:\\/(\\S*))?$`);
+  const ignoredPeers = new RegExp(`^${escapeForRegex(`${prefix}${suffix}`)}peer(?:\\/(\\S*))?$`);
+
+  const customComponentClasses = detectComponentClasses
+    ? getCustomComponentClasses({ configPath: tailwindConfig, cwd: ctx.cwd })
+    : [];
 
   for(const literal of literals){
 
-    const { ignore, tailwindConfig } = getOptions(ctx);
-
     const classes = splitClasses(literal.content);
 
-    const [unregisteredClasses, warnings] = getUnregisteredClasses({ classes, configPath: tailwindConfig, cwd: ctx.cwd });
-
-    const unregisteredClassesWarnings = warnings.map(warning => ({ ...warning, url: DOCUMENTATION_URL }));
+    const { unregisteredClasses, warnings } = getUnregisteredClasses({ classes, configPath: tailwindConfig, cwd: ctx.cwd });
 
     if(unregisteredClasses.length === 0){
       continue;
     }
 
     for(const unregisteredClass of unregisteredClasses){
-      if(ignore.some(ignoredClass => unregisteredClass.match(ignoredClass))){
+      if(
+        ignore.some(ignoredClass => unregisteredClass.match(ignoredClass)) ||
+        customComponentClasses.includes(unregisteredClass) ||
+        unregisteredClass.match(ignoredGroups) ||
+        unregisteredClass.match(ignoredPeers)
+      ){
         continue;
       }
 
@@ -126,7 +144,11 @@ function lintLiterals(ctx: Rule.RuleContext, literals: Literal[]) {
           unregistered: display(unregisteredClass)
         },
         loc: getExactClassLocation(literal, unregisteredClass),
-        message: augmentMessageWithWarnings("Unregistered class detected: {{ unregistered }}", unregisteredClassesWarnings)
+        message: augmentMessageWithWarnings(
+          "Unregistered class detected: {{ unregistered }}",
+          DOCUMENTATION_URL,
+          warnings
+        )
       });
     }
 
@@ -140,9 +162,11 @@ export function getOptions(ctx: Rule.RuleContext) {
   const common = getCommonOptions(ctx);
 
   const ignore = options.ignore ?? defaultOptions.ignore;
+  const detectComponentClasses = options.detectComponentClasses ?? defaultOptions.detectComponentClasses;
 
   return {
     ...common,
+    detectComponentClasses,
     ignore
   };